# Cleanup and make CargoDirectories::new fast path faster