Simplifies the implementation, makes it more robust (traverses all parent dirs instead of just 1), and makes the fast path faster.

The previous implementation of CargoDirectories::new took 110us to complete on wgpu.
With this change it takes 7us on wgpu.

In comparison the slow path takes 12ms (I'm on a different machine now), so we certainly dont care about initially waiting 7us there.